Split and Refactotr the code.
