+++ Frontend/index.php
@@ -4,15 +4,11 @@
         <meta charset="UTF-8">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <script src="https://cdn.jsdelivr.net/npm/@tailwindcss/browser@4"></script>
-        <!-- <script type="module" src="./Js/main.js"></script>
-        <script type="module" src="./Js/pageRouting.js"></script> -->
-        <script type="module" src="./Js/customElements.js"></script>
         <title>Laboratory Scheduling System</title>
     </head>
     <body class="w-full h-screen bg-[url('./resources/img/Wallpaper.jpg')] bg-cover bg-center bg-no-repeat backdrop-blur-xs "  id="index-content">
         <div class="w-full h-screen overflow-y-scroll pb-24">
             <?php include __DIR__ . '/Components/NavigationBar.php';?>
-            <!-- <page-router></page-router> -->
             <main class=" w-full">
                 <section class="flex flex-col w-full justify-between">
                     <article class="w-2/3 p-2 pt-24">
@@ -31,7 +27,6 @@
                 </section>
             </main>
             <?php include __DIR__ . '/Components/FooterBar.php';?>
-            <!-- <page-footer></page-footer> -->
         </div>
     </body>
     <script type="module" src="./JS/main.js"></script>
